Dockerized app

+++ backend/src/connections/db.ts
@@ -1,31 +1,51 @@
-import { createConnection, getConnection } from 'typeorm';
-import { dbConfig } from '../config';
+import { createConnection, getConnectionManager } from 'typeorm';
+/**
+ * For some reason this import has to be directly to `../config/typeorm` rather than `../config`
+ * Otherwise it's undefined
+ * ???
+ */
+import { dbConfig } from '../config/typeorm';
 
 export const dbConnection = async (dropSchema?: boolean, synchronize?: boolean) => {
-  if (!dropSchema) {
-    if (dbConfig.dropSchema) {
-      dropSchema = dbConfig.dropSchema;
-    } else {
-      dropSchema = false;
+  const connectionManager = getConnectionManager();
+  if (!connectionManager.has('default')) {
+    if (!dropSchema) {
+      if (dbConfig.dropSchema) {
+        dropSchema = dbConfig.dropSchema;
+      } else {
+        dropSchema = false;
+      }
     }
-  }
 
-  if (!synchronize) {
-    if (dbConfig.synchronize) {
-      synchronize = dbConfig.synchronize;
-    } else {
-      synchronize = false;
+    if (!synchronize) {
+      if (dbConfig.synchronize) {
+        synchronize = dbConfig.synchronize;
+      } else {
+        synchronize = false;
+      }
     }
-  }
 
-  try {
-    return getConnection('default');
-  } catch (e) {
-    return await createConnection({
-      ...dbConfig,
-      name: 'default',
-      dropSchema,
-      synchronize
-    });
+    let retries = 5;
+
+    while (retries) {
+      try {
+        return createConnection({
+          ...dbConfig,
+          name: 'default',
+          dropSchema,
+          synchronize
+        });
+      } catch (e) {
+        console.log(e);
+        retries -= 1;
+        console.log(`Retries left: ${retries}`);
+        // wait 5 seconds
+        await new Promise(res => setTimeout(res, 5000));
+      }
+    }
+
+    return undefined;
+  } else {
+    return connectionManager.get('default');
   }
 };
